requirements.txt
requests~=2.31.0
pytest~=7.4.0
pydantic~=2.4.2
typing_extensions~=4.8.0
python-dotenv~=1.0.0
Faker~=19.11.0
MarkupSafe~=2.1.5
jinja2~=3.1.4
pytest-html~=4.1.1
pytest-metadata~=3.1.1
